# Remove the last decoder allocation