--- a/zstd/framedec.go
+++ b/zstd/framedec.go
@@ -40,7 +40,7 @@ type frameDec struct {
 	rawInput byteBuffer
 
 	// Optional byte buffer that can be reused.
-	bBuf *byteBuf
+	bBuf byteBuf
 
 	// asyncRunning indicates whether the async routine processes input on 'decoding'.
 	asyncRunning   bool