```
benchmark                                            old ns/op     new ns/op     delta
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12          608499        610005        +0.25%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12      136501        137699        +0.88%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12       1967989       1944000       -1.22%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12         1448010       1444000       -0.28%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12       504333        504333        +0.00%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12        650494        652504        +0.31%
BenchmarkDecoder_DecodeAll/html_x_4.zst-12           267000        267199        +0.07%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12     24840         24659         -0.73%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12     9594          9639          +0.47%
BenchmarkDecoder_DecodeAll/urls.10K.zst-12           1662010       1670000       +0.48%
BenchmarkDecoder_DecodeAll/html.zst-12               156400        155601        -0.51%

benchmark                                            old MB/s     new MB/s     speedup
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12          302.91       302.16       1.00x
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12      868.77       861.21       0.99x
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12       244.85       247.87       1.01x
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12         294.72       295.54       1.00x
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12       248.21       248.21       1.00x
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12        233.81       233.09       1.00x
BenchmarkDecoder_DecodeAll/html_x_4.zst-12           1534.08      1532.93      1.00x
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12     4122.35      4152.47      1.01x
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12     12828.87     12769.05     1.00x
BenchmarkDecoder_DecodeAll/urls.10K.zst-12           422.43       420.41       1.00x
BenchmarkDecoder_DecodeAll/html.zst-12               654.73       658.09       1.01x

benchmark                                            old allocs     new allocs     delta
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12          1              0              -100.00%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12      1              0              -100.00%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12       1              0              -100.00%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12         1              0              -100.00%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12       1              0              -100.00%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12        1              0              -100.00%
BenchmarkDecoder_DecodeAll/html_x_4.zst-12           1              0              -100.00%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12     1              0              -100.00%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12     1              0              -100.00%
BenchmarkDecoder_DecodeAll/urls.10K.zst-12           1              0              -100.00%
BenchmarkDecoder_DecodeAll/html.zst-12               1              0              -100.00%

benchmark                                            old bytes     new bytes     delta
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12          32            0             -100.00%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12      32            0             -100.00%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12       32            0             -100.00%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12         32            0             -100.00%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12       32            0             -100.00%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12        32            0             -100.00%
BenchmarkDecoder_DecodeAll/html_x_4.zst-12           32            0             -100.00%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12     32            0             -100.00%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12     32            0             -100.00%
BenchmarkDecoder_DecodeAll/urls.10K.zst-12           32            0             -100.00%
BenchmarkDecoder_DecodeAll/html.zst-12               32            0             -100.00%
```